no entra en el if de eliminar mi publicacion

+++ src/screens/Home.js
@@ -21,7 +21,7 @@ import UserContext from "../context/UserContext";
 import { useNavigation } from "@react-navigation/native";
 
 const Home = () => {
-  const IP = "10.144.1.15";
+  const IP = "10.152.2.130";
   const { token } = useContext(TokenContext);
   const { user } = useContext(UserContext);
   const [likesFromUser, setLikesFromUser] = useState([]);
@@ -209,19 +209,27 @@ const Home = () => {
     await getAllDataFromPublication();
   };
 
-  {/*funcion iaacob */}
+  {/*eliminar publicacion */ }
   const deletePublication = async (id) => {
-    console.log("holi estoy en la funcion")
-    await axios.delete(`http://${IP}:4000/publicaciones/${id}` , {
-      headers: {
-        "Content-Type": "application/json",
-        Authorization: `bearer ${token}`,
-        "X-Requested-With": "XMLHttpRequest",
-      },
-    });
-    await obtenerLikesDelUser(user);
-    await obtenerDislikesDelUser(user);
-    await getAllDataFromPublication();
+    if (id == logedUser.Id) {
+
+      console.log("estoy en el if: ", id, logedUser.Id)
+
+      await axios.delete(`http://${IP}:4000/publicaciones/${id}`, {
+        headers: {
+          "Content-Type": "application/json",
+          Authorization: `bearer ${token}`,
+          "X-Requested-With": "XMLHttpRequest",
+        },
+
+      });
+      await obtenerLikesDelUser(user);
+      await obtenerDislikesDelUser(user);
+      await getAllDataFromPublication();
+
+    } else {
+      alert("no se puede eliminar")
+    }
   }
 
   const onRefresh = async () => {
@@ -303,9 +311,8 @@ const Home = () => {
                     >
                       {item.occupation}
                     </Text>
-                    {/* boton iaacob */}
                     <Button
-                      onPress={ async ()=>deletePublication(item.Id)}
+                      onPress={async () => deletePublication(item.id)}
                       title="Delete"
                       color="#841584"
                       accessibilityLabel="Learn more about this purple button"
